+++ CHANGELOG.md
@@ -24,6 +24,8 @@ The following changes have been completed, and will be included in the next tagg
 ### New
 
 * Basic Types
+  - added `Immutable`
+  - added `DeepImmutable`
   - added `Primitive`
   - added `PrimitiveOrFunction`
   - added `PrimitiveOrUndefined`